move template into types
